less logging

+++ ApplicationBase/test/org/mcv/app/TestApplication.java
@@ -15,15 +15,17 @@ public class TestApplication {
 
 	@BeforeClass
 	public static void setupClass() {
-		clearDirectory(new File("./logs"));		
+		clearDirectory(new File("./logs"));
 	}
 
 	private static void clearDirectory(File dir) {
-		for(File f : dir.listFiles()) {
-			if(f.isDirectory()) {
-				clearDirectory(f);
+		if (dir.exists()) {
+			for (File f : dir.listFiles()) {
+				if (f.isDirectory()) {
+					clearDirectory(f);
+				}
+				f.delete();
 			}
-			f.delete();
 		}
 	}
 
@@ -32,7 +34,7 @@ public class TestApplication {
 		app.clear();
 		return app;
 	}
-	
+
 	/* app should initialize DB */
 	@Test
 	public void testApplicationDB() {
@@ -40,7 +42,7 @@ public class TestApplication {
 		Base log = app.create();
 		try {
 			assertTrue(app.getDb() != null);
-		} catch(Exception e) {
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
@@ -58,23 +60,26 @@ public class TestApplication {
 			assertTrue(app.getDb() != null);
 			// create and persist
 			Base base = app.create("base", Base.class);
-			for(LogEntry entry : app.log.getLogs()) {
+			for (LogEntry entry : app.log.getLogs()) {
 				System.out.println(entry);
 			}
 			assertTrue(base != null);
-			//System.out.println("created: " + base.toString());
+			// System.out.println("created: " + base.toString());
 			// retrieve
 			Base copy = app.create("base", Base.class);
-			//System.out.println("retrieved: " + copy.toString());
+			// System.out.println("retrieved: " + copy.toString());
 			assertTrue(base.equals(copy));
-		} catch(Exception e) {
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
 		}
 	}
 
-	/* app should be able to create and persist and retrieve Base subclass objects */
+	/*
+	 * app should be able to create and persist and retrieve Base subclass
+	 * objects
+	 */
 	@Test
 	public void testApplicationBaseSubCreate() {
 		Application app = mkApp();
@@ -84,12 +89,12 @@ public class TestApplication {
 			// create and persist
 			BaseSub base = app.create("base", BaseSub.class);
 			assertTrue(base != null);
-			//System.out.println(base.toString());
+			// System.out.println(base.toString());
 			// retrieve
 			BaseSub copy = app.create("base", BaseSub.class);
-			//System.out.println(copy.toString());
+			// System.out.println(copy.toString());
 			assertTrue(base.equals(copy));
-		} catch(Exception e) {
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
@@ -106,27 +111,30 @@ public class TestApplication {
 			// create and persist
 			Base base = app.create("base", Base.class);
 			assertTrue(base != null);
-			//System.out.println("created: " + base.toString());
+			// System.out.println("created: " + base.toString());
 			base.delete();
-			//System.out.println("deleted: " + base.toString());
+			// System.out.println("deleted: " + base.toString());
 			// retrieve
 			Base copy = app.create("base", Base.class);
-			if(copy != null) {
+			if (copy != null) {
 				System.out.println("deleted: " + copy.toString());
 			}
 			assertTrue(copy == null);
 			base.undelete();
 			copy = app.create("base", Base.class);
 			assertTrue(copy.equals(base));
-			
-		} catch(Exception e) {
+
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
 		}
 	}
 
-	/* app should be able to retrieve current version of changed subclass objects */
+	/*
+	 * app should be able to retrieve current version of changed subclass
+	 * objects
+	 */
 	@Test
 	public void testApplicationBaseChanged() {
 		Application app = mkApp();
@@ -136,14 +144,14 @@ public class TestApplication {
 			// create and persist
 			BaseSub base = app.create("base", BaseSub.class);
 			assertTrue(base != null);
-			//System.out.println("created: " + base.toString());
+			// System.out.println("created: " + base.toString());
 			base.setMsg("Changed");
-			//System.out.println("deleted: " + base.toString());
+			// System.out.println("deleted: " + base.toString());
 			// retrieve
 			BaseSub copy = app.create("base", BaseSub.class);
 			assertTrue(copy.equals(base));
 
-		} catch(Exception e) {
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
@@ -158,35 +166,35 @@ public class TestApplication {
 		try {
 			assertTrue(app.getDb() != null);
 			List<BaseSub> list1 = new ArrayList<>();
-			for(int i=0; i < 10; i++) {
+			for (int i = 0; i < 10; i++) {
 				// create and persist
 				BaseSub base = app.create("base" + i, BaseSub.class);
 				assertTrue(base != null);
 				list1.add(base);
 			}
-			for(int i=0; i < 10; i++) {
+			for (int i = 0; i < 10; i++) {
 				// change and persist
 				BaseSub base = list1.get(i);
-				base.setMsg("Changed"+i);
+				base.setMsg("Changed" + i);
 			}
 			// retrieve
 			List<BaseSub> list2 = app.getList(BaseSub.class);
 
 			assertEquals(10, list2.size());
-			for(int i=0; i < 10; i++) {
+			for (int i = 0; i < 10; i++) {
 				BaseSub base1 = list1.get(i);
 				System.out.println(base1);
 				BaseSub base2 = list2.get(i);
 				System.out.println(base2);
-				assertTrue(base1.equals(base2));				
+				assertTrue(base1.equals(base2));
 			}
-		} catch(Exception e) {
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
 		}
 	}
-	
+
 	@Test
 	public void testGetLogs() {
 		Application app = mkApp();
@@ -206,14 +214,14 @@ public class TestApplication {
 			base.error(new Exception("Really bad"), "This is an error message");
 			assertTrue(base.getVersion() == 5);
 			List<LogEntry> logs = base.getLogs();
-			for(LogEntry logentry : logs) {
+			for (LogEntry logentry : logs) {
 				System.out.println(logentry);
 			}
-		} catch(Exception e) {
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
-		}		
+		}
 	}
 
 	@Test
@@ -237,14 +245,14 @@ public class TestApplication {
 			base.error(new Exception("Really bad"), "This is an error message");
 			assertTrue(base.getVersion() == 2);
 			List<LogEntry> logs = base.getLogs();
-			for(LogEntry logentry : logs) {
+			for (LogEntry logentry : logs) {
 				System.out.println(logentry);
 			}
-		} catch(Exception e) {
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
-		}		
+		}
 	}
 
 	@Test
@@ -260,11 +268,11 @@ public class TestApplication {
 			base.store();
 			// no effect
 			assertTrue(base.getVersion() == 2);
-		} catch(Exception e) {
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
-		}		
+		}
 	}
 
 	@Test
@@ -273,31 +281,31 @@ public class TestApplication {
 		Base log = app.create();
 		try {
 			assertTrue(app.getDb() != null);
-			
+
 			// create version 1
 			BaseSub base = app.create("base", BaseSub.class);
 			assertTrue(base != null);
-			
+
 			// create version 2
 			base.setMsg("Changed");
-			
+
 			// create version 3
 			base.setMsg("Changed again");
-			
+
 			// create version 4
 			base.setMsg("Changed yet again");
-			
+
 			// create version 5
 			base.setMsg("Keep on changing");
-			
+
 			assertTrue(base.getVersion() == 5);
-			
+
 			base.undo();
 			assertTrue(base.getVersion() == 4);
 			assertEquals("Changed yet again", base.getMsg());
 
 			base.undo();
-			assertTrue(""+base.getVersion(), base.getVersion() == 3);
+			assertTrue("" + base.getVersion(), base.getVersion() == 3);
 			assertEquals("Changed again", base.getMsg());
 
 			base.undo();
@@ -311,11 +319,11 @@ public class TestApplication {
 			base.undo();
 			assertTrue(base.getVersion() == 1);
 			assertEquals("Hello", base.getMsg());
-			
+
 			base.redo();
 			assertTrue(base.getVersion() == 2);
 			assertEquals("Changed", base.getMsg());
-			
+
 			base.redo();
 			assertTrue(base.getVersion() == 3);
 			assertEquals("Changed again", base.getMsg());
@@ -327,7 +335,7 @@ public class TestApplication {
 			base.redo();
 			assertTrue(base.getVersion() == 5);
 			assertEquals("Keep on changing", base.getMsg());
-			
+
 			// now it gets complicated
 			base.undo();
 			assertTrue(base.getVersion() == 4);
@@ -344,7 +352,7 @@ public class TestApplication {
 			base.redo();
 			assertTrue(base.getVersion() == 2);
 			assertEquals("Changed", base.getMsg());
-			
+
 			base.setMsg("New branch");
 			assertTrue(base.getVersion() == 6);
 			assertEquals("New branch", base.getMsg());
@@ -355,15 +363,15 @@ public class TestApplication {
 
 			// we cannot redo
 			assertTrue(!base.canRedo());
-			
+
 			List<Long> options = base.redoOptions();
 			assertTrue(options.get(0) == 3);
 			assertTrue(options.get(1) == 6);
-			
+
 			base.redo(options.get(0));
 			assertTrue(base.getVersion() == 3);
 			assertEquals("Changed again", base.getMsg());
-			
+
 			base.undo();
 			base.redo(6);
 			assertTrue(base.getVersion() == 6);
@@ -372,9 +380,8 @@ public class TestApplication {
 			base.undo();
 			assertTrue(base.getVersion() == 1);
 			assertEquals("Hello", base.getMsg());
-			
-			
-		} catch(Exception e) {
+
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
@@ -397,28 +404,27 @@ public class TestApplication {
 			assertTrue(base.getVersion() == 5);
 			List<BaseSub> versions = base.getVersions();
 			assertEquals(5, versions.size());
-			int i=1;
-			for(BaseSub version : versions) {
+			int i = 1;
+			for (BaseSub version : versions) {
 				System.out.println(version);
 				assertEquals(i++, version.getVersion());
 				assertTrue(!version.isCurrent());
 			}
-			
+
 			base.select(versions.get(2));
-			assertTrue(""+base.getVersion(), base.getVersion() == 3);
+			assertTrue("" + base.getVersion(), base.getVersion() == 3);
 			assertEquals("Changed again", base.getMsg());
 			// select is not undoable!
 			base.undo();
 			assertTrue(base.getVersion() == 2);
 			assertEquals("Changed", base.getMsg());
-			
-		} catch(Exception e) {
+
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
-		}		
+		}
 	}
-	
 
 	// app.getAllLogs
 	@Test
@@ -433,20 +439,20 @@ public class TestApplication {
 			base.setMsg("Changed again");
 			base.setMsg("Changed yet again");
 			base.setMsg("Keep on changing");
-			
+
 			List<LogEntry> logs = app.getAllLogs();
-			for(LogEntry log : logs) {
+			for (LogEntry log : logs) {
 				/* :) */
 				app.log.log(log);
 			}
 			assertTrue(logs.size() > 0);
-		} catch(Exception e) {
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			logger.error(t.toString(), t);
 			fail(t.toString());
-		}		
+		}
 	}
-	
+
 	@Test
 	public void testAutoDeclare() {
 		Application app = mkApp();
@@ -456,11 +462,11 @@ public class TestApplication {
 			BaseSub base = app.create();
 			assertEquals("base", base.getName());
 			assertEquals(BaseSub.class, base.getClazz());
-		} catch(Exception e) {
+		} catch (Exception e) {
 			Throwable t = WrapperException.unwrap(e);
 			log.error(t.toString(), t);
 			fail(t.toString());
-		}		
+		}
 	}
 
 }
